update_0402

+++ DirectX11_Engine/GameClient/CFlipbookRender.cpp
@@ -105,7 +105,7 @@ void CFlipbookRender::FinalTick()
 		m_AccTime -= fLmit;
 		++m_CurSprite;
 
-		if (m_vecFlipbook[m_CurFlipbook]->GetSpriteCount() <= m_CurSprite)
+		if (m_vecFlipbook[m_CurFlipbook]->GetSpriteCount() <= (UINT)m_CurSprite)
 		{
 			m_Finish = true;
 			--m_CurSprite;
